% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lib_assets}
\alias{add_lib_assets}
\title{Add library assets (useful in knitr documents)}
\usage{
add_lib_assets(lib, cdn = F, standalone = F)
}
\description{
Add library assets (useful in knitr documents)
}

