% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highmaps_Choropleth.R
\name{highmapsChoropleth}
\alias{highmapsChoropleth}
\title{Plot Interactive map using Highmaps}
\usage{
highmapsChoropleth(data, value, key, time = NULL, title = "",
  scope = "France_regions", col = if (is.character(data[, value]) |
  is.factor(data[, value])) {     c("#2f7ed8", "#0d233a", "#8bbc21", "#910000",
  "#1aadce", "#492970", "#f28f43", "#77a1e5", "#c42525", "#a6c96a") } else {    
  c("#FAFAFA", "#0A0A2A") })
}
\arguments{
\item{data}{: data.frame}

\item{value}{: column name of quantitative/qualitative value}

\item{key}{: column name of key value to identify regions / countries. We have to use "hc-key" reference for identification.
See data(Europe), data(UE), data(France_regions) and data(France_regions_dom) for reference table.}

\item{time}{: Create an animated map ? Column name of time value, default to NULL (no time)}

\item{title}{: title, default to ""}

\item{scope}{: scope of the map, one of "Europe", "UE", "France_regions" (default) or "France_regions_dom"}

\item{col}{: vector of colors (html code rather). Two colors for quantitative variable, n colors for qualitative variable}
}
\value{
a Highmaps object
}
\description{
Plot Interactive map using Highmaps. France (regions) with or without dom, Europe and UE (countries)
}
\details{
Plot Interactive map using Highmaps. France (regions) with or without dom, Europe and UE (countries)
}
\examples{
#------------   UE EXAMPLE, quantitative, not time ------------------#

data(UE)
head(UE)
UE$value <- rnorm(nrow(UE), 50,100)

res <- highmapsChoropleth(UE, value ="value", key = "hc-key", title = "Exemple UE",
 scope = "UE")
res

# changing colors
res <- highmapsChoropleth(UE, value ="value", key = "hc-key", title = "Exemple UE",
 scope = "UE",
 col = c("#FBEFFB","#380B61"))
res

#------------   EUROPE EXAMPLE, quantitative, with time ------------------#

data(Europe)
head(Europe)

Europe <- rbind(cbind(Europe, time = "01-01-2010"), cbind(Europe, time = "02-01-2010"),
 cbind(Europe, time = "03-01-2010"), cbind(Europe, time = "04-01-2010"))

Europe$value <- rnorm(nrow(Europe),50,100)

# add NA for France and delete Spain to test
Europe$value[Europe$name=="France"] <- NA
Europe <- Europe[-which(Europe$name=="Spain"), ]

res <- highmapsChoropleth(Europe, value ="value", key = "hc-key", time = "time",
 title = "Exemple Europe", scope = "Europe")
res

#------------   FRANCE EXAMPLE, qualitative, not time ------------------#

data(France_regions)
head(France_regions)
France_regions$value <- as.factor(round(runif(nrow(France_regions))*4,0))

# NA and Suppress row for test
France_regions$value[ which(France_regions$name=="Bretagne")] <- NA
France_regions <- France_regions[-which(France_regions$name=="Basse-Normandie"),]

colnames(France_regions)[1] = "key"
res <- highmapsChoropleth(France_regions, value ="value", key = "key",
 title = "Exemple France", scope = "France_regions")
res

# changing colors

res <- highmapsChoropleth(France_regions, value ="value", key = "hc-key",
  scope = "France_regions", col = c("red","green","blue","black", "yellow"))
res

#------------   FRANCE with DOM EXAMPLE, qualitative, time ------------------#

data(France_regions_dom)
head(France_regions_dom)

France_regions_dom <- rbind(cbind(France_regions_dom, time = "first"),
 cbind(France_regions_dom, time = "second"),
 cbind(France_regions_dom, time = "third"),
 cbind(France_regions_dom, time = "fourth"))

France_regions_dom$value <- as.factor(round(runif(nrow(France_regions_dom))*4,0))

# Na and Suppress row for test
France_regions_dom <- France_regions_dom[-which(France_regions_dom$time=="first" &
 France_regions_dom$value == "4"),]
France_regions_dom[30,"value"] <- NA

res <- highmapsChoropleth(France_regions_dom, value ="value", key = "hc-key", time = "time",
  title = "Exemple France", scope = "France_regions_dom")
res

save.graphic(res,file = "mychoropleth.html")
}

