% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Highcharts.R
\name{hPlot}
\alias{hPlot}
\title{Highcharts Plot}
\usage{
hPlot(..., radius = 3, title = NULL, subtitle = NULL, group.na = NULL)
}
\arguments{
\item{...}{see getLayer(...)}

\item{radius}{circle size}

\item{title}{chart title}

\item{subtitle}{chart subttitle}

\item{group.na}{replace NA's with chosen value in the group column, or else these observations will be removed}
}
\description{
...
}

