% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot time series}
\usage{
plotTimeSeries(data, ind.date, ind.time = NULL, ind.curve = if
  (!is.null(data)) {     3:ncol(data) } else {     NULL }, title = "",
  subtitle = "", xlab = "", ylab = "", zoomType = "x", lwd = 1,
  cex = 1, type = "line", suffixe.value = "", approximation = "average",
  dash.style = c("solid", "shortdash", "shortdot", "shortdashdot",
  "shortdashdotdot", "dot", "dash", "longlash", "dashdot", "longdashdot",
  "longdashdotdot"), enabled.marker = TRUE, marker = c("circle", "square",
  "diamond", "triangle", "triangle-down"), width = "100\%", height = 500,
  tz = "UTC", plot = FALSE, col = c("#2f7ed8", "#0d233a", "#8bbc21",
  "#910000", "#1aadce", "#492970", "#f28f43", "#77a1e5", "#c42525", "#a6c96a"),
  legend.enabled = TRUE, legend.layout = "vertical",
  legend.halign = "left", legend.valign = "middle")
}
\arguments{
\item{data}{: data.frame}

\item{ind.date}{: indice of data column}

\item{ind.time}{: indice of time column, if NULL, on take date column}

\item{ind.curve}{: time series to plot. Default to if(!is.null(data)){3:ncol(data)}else{NULL}}

\item{title}{: Title. Optional ,"" by default}

\item{subtitle}{: Subtitle. Optional ,"" by default}

\item{xlab}{: x label. Optional ,"" by default}

\item{ylab}{: y label. Optional ,"" by default}

\item{zoomType}{: zoomType one of "x" (default), "y", "xy", or ""}

\item{lwd}{: line width. Optional, 1 by default}

\item{cex}{: marker size. Optional, 1 by default}

\item{type}{: type of graphic. one of "line" (default), "spline", or "column"}

\item{suffixe.value}{: add a suffixe on y legend ? "" by default}

\item{approximation}{: approximation used during aggregation. one of "average" (default), "high", "low", or "sum"}

\item{dash.style}{: like lty R, value(s) in  c("solid","shortdash","shortdot","shortdashdot","shortdashdotdot","dot","dash","longlash","dashdot","longdashdot", "longdashdotdot")}

\item{enabled.marker}{: See arkers on graphic ? TRUE by default}

\item{marker}{: marker type, value(s) in c("circle", "square", "diamond", "triangle","triangle-down")}

\item{height}{: height of graphic. Default to 500}

\item{tz}{: time zone of date/time, used in graphic transformation. "UTC" by default}

\item{plot}{: If assign graphic to an R object, plot during calculation ? FALSE by default}

\item{col}{: color(s) of curve(s). can be some R colors or html colors. Default to c('#2f7ed8', '#0d233a', '#8bbc21', '#910000', '#1aadce',
'#492970', '#f28f43', '#77a1e5', '#c42525', '#a6c96a')}

\item{legend.enabled}{: Add a legend ? Default to TRUE}

\item{legend.layout}{: layout legend. "vertical" (Default) or "horizontal"}

\item{legend.halign}{: horizontal alignment "left" (Default), "center", or "right"}

\item{legend.valign}{: vertical alignment "middle" (Default), "top" or "bottom"}
}
\value{
a Highcharts object
}
\description{
Plot time series
}
\details{
Plot time series
}
\examples{
data(timeSeriesData)
head(timeSeriesData)


# simple plot
data = timeSeriesData[1:500,]
plotTimeSeries(data = data, ind.date = 1, ind.time = 2)

# customize
plotTimeSeries(data = data, ind.date = 1, ind.time = 2, cex = 3, marker = "diamond", type = "line", height = 600,
                title = "My Time Series",subtitle = "Customize !", xlab = "Time", ylab = "Puissance",
                suffixe.value = "W", col = "green", legend.halign = "center", legend.valign= "bottom")

# plot, keep, and save
graph <- plotTimeSeries(data = data, ind.date = 1, ind.time = 2, plot = TRUE)
save.graphic(graph,"mysavetimeseries.html")

# keep, and save, no plot
graph <- plotTimeSeries(data = data, ind.date = 1, ind.time = 2, col = "green")
save.graphic(graph,"mysavetimeseries2.html")

# plot two daily time series
don = data.frame(date = seq(as.Date("1999/1/1"), as.Date("1999/12/31"), "days"), Value1 = rnorm(365,50,10), Value2 = rnorm(365,50,10))
head(don)

plotTimeSeries(don, ind.date = 1, ind.curve = c(2:3), type = "spline", suffixe.value = "W", col = c("green","red"), plot = FALSE)
}

