% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotWindRose}
\alias{plotWindRose}
\title{Plot wind rose}
\usage{
plotWindRose(data, point.grid, data.add = NULL, point.grid.add = NULL,
  range.direction = c(0, 360), ori.direction = c("N", "E"), type = "rose",
  breaks.vitesse = if (type == "rose") {     c(0, 0.5, 2, 4, 6, 8, 10, Inf) }
  else {     c(0.25, 0.5, 0.75) }, title = "Wind Rose",
  vitesse.name = "VITESSE", direction.name = "DIRECTION",
  col = c("#000000", "#0B4C5F", "#086A87", "#01A9DB", "#2ECCFA", "#81DAF5",
  "#CEECF5", "#EFF8FB"), plot = FALSE, height = 450)
}
\arguments{
\item{data}{: data.frame, with vitesse.name+point.grid and direction.name+point.grid columns. See example}

\item{point.grid}{: vector of numeric point of grid for first serie}

\item{data.add}{: If wanted to add a second serie on another data, second data.frame. NULL by default}

\item{point.grid.add}{: If wanted to add a second serie, vector of numeric point of grid. NULL by default. If not null and data.add is null, second serie is based on data}

\item{range.direction}{: range of direction values. c(0,360) by default.}

\item{ori.direction}{: Direction for the minimum value, and for the first 1/4 circle. Can be two of "W", "S", "N", "E", .... c("N","E") by default. N = 0, and E = 90}

\item{type}{: type of graphic & transformation. "rose" for classic wind rose, "quantile" to apply quantile per direction.}

\item{breaks.vitesse}{: breaks for direction transformation. Default to = c(0,0.5,2,4,6,8,10,Inf) for "cut" and c(0,0.05,0.25,0.5,0.75,0.95,1) for"quantile"}

\item{title}{: title, default to "Wind Rose"}

\item{vitesse.name}{: name of vitesse columns in data. Default to "VITESSE"}

\item{direction.name}{: name of direction columns in data. Default to "DIRECTION"}

\item{col}{: vector of colors (html code rather). From strong wind to low wind. Default to c('#000000', '#0B4C5F', '#086A87','#01A9DB', '#2ECCFA', '#81DAF5', '#CEECF5', '#EFF8FB')}

\item{plot}{: If assign graphic to an R object, plot during calculation ? FALSE by default}

\item{height}{: height of output, default to 450}
}
\value{
a Highcharts object
}
\description{
Plot wind rose
}
\details{
Plot wind rose
}
\note{
Copyright 2014 RTE
}
\examples{
require(highchartsUtils)
data(windRoseData)
head(windRoseData)

# one serie, wind Rose (type = "rose", default value)
plotWindRose(data = windRoseData, point.grid = 1, title ="Wind Rose")

# change direction orientation
plotWindRose(data = windRoseData, point.grid = 1, ori.direction = c("N","W"))

# one serie, quantile
plotWindRose(data = windRoseData, point.grid = 1, type  = "quantile", title ="Wind Rose - quantile")

# change colors, keeping result
res <- plotWindRose(data = windRoseData, point.grid = 1, title ="Wind Rose",
 col = c("#AC58FA","#8A0808", "#0B2161","#DF0174","#D0FA58"))
res

plotWindRose(data = windRoseData, point.grid = 1, title ="Wind Rose",
 col = c("red","blue","green","yellow"))

# type = "rose", change breaks
plotWindRose(data = windRoseData, point.grid = 1, title ="Wind Rose - Change break",
 breaks.vitesse = c(0,10,20,30,Inf))

# two series, on same data
plotWindRose(data = windRoseData, point.grid = 1, point.grid.add = 2:3,
 breaks.vitesse = c(0,2,5,10,15,20,Inf))

# two series - two dataset, quantile per Direction, changing quantile
res <- plotWindRose(data = windRoseData, point.grid = 1, data.add = windRoseData[1:1000,], point.grid.add = 1,
 type = "quantile", breaks.vitesse = c(0.3, 0.5,0.8))

res

save.graphic(res,file = "Package_libraries_to_file.html")
save.graphic(res,file = "Internet_libraries_to_file.html", use.internet.library = TRUE)
save.graphic(res,file = "Package_libraries_linking_file.html", write.library.to.file = FALSE)
save.graphic(res,file = "Internet_libraries_linking_file.html", use.internet.library = TRUE, write.library.to.file = FALSE)
}

