% Generated by roxygen2 (4.0.2): do not edit by hand
\name{renderHighcharts}
\alias{renderHighcharts}
\title{renderHighcharts (use with Shiny)}
\usage{
renderHighcharts(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a chart object}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is expr a quoted expression (with \code{quote()})? This is
 useful if you want to save an expression in a variable.}
}
\description{
Use as Shiny output. First, use \code{renderHighcharts} in \code{server.R}
to assign the chart object to an Shiny output. Then create an highchartsOutput
with the same name in #' \code{ui.R}. \code{highchartsOutput} is currently just an
alias for \code{htmlOutput}.
}
\author{
Thomas Reinholdsson, Ramnath Vaidyanathan
}

