% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/save_graphic.R
\name{save.graphic}
\alias{save.graphic}
\title{Save graphic in a .html file}
\usage{
save.graphic(graphic, file = "NewGraphic.html",
  use.internet.library = FALSE, write.library.to.file = TRUE)
}
\arguments{
\item{file}{: html file. "NewGraphic.html" by default}

\item{use.internet.library}{: Use internet online libray ? Default to FALSE, so use package library}

\item{write.library.to.file}{: write.library.to.file ? Default to TRUE. Writing library to file,
graphic can be open by any person on any machine.}

\item{graphics}{: graphic to save}
}
\value{
a Highcharts object
}
\description{
Save graphic in a .html file
}
\details{
Save graphic in a .html file
}
\note{
Copyright 2014 RTE
}
\examples{
data(windRoseData)
head(windRoseData)

plotWindRose(data = windRoseData, point.grid = 1)

res <- plotWindRose(data = windRoseData, point.grid = 1, title ="Wind Rose - grid 1",
 breaks.vitesse = c(0,2,5,10,15,20,30,Inf), plot = FALSE)

res

save.graphic(res,file = "Package_libraries_to_file.html")
save.graphic(res,file = "Internet_libraries_to_file.html", use.internet.library = TRUE)
save.graphic(res,file = "Package_libraries_linking_file.html", write.library.to.file = FALSE)
save.graphic(res,file = "Internet_libraries_linking_file.html", use.internet.library = TRUE, write.library.to.file = FALSE)
}

