% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/toJSON.R
\name{toGeoJSON}
\alias{toGeoJSON}
\title{Convert a list to a GeoJSON compatible list.}
\usage{
toGeoJSON(list_, lat = "latitude", lon = "longitude")
}
\description{
Convert a list to a GeoJSON compatible list.
}

