% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highchartsOutput.R
\name{highchartsOutput}
\alias{highchartsOutput}
\title{Use as Shiny output. First, use \code{renderHighcharts} in \code{server.R} to assign
the chart object to an Shiny output. Then create an chartOutput with the same name in #'
\code{ui.R}.}
\usage{
highchartsOutput(outputId, lib = "highcharts", package = "highchartsUtils",
  add_lib = TRUE)
}
\arguments{
\item{outputId}{output variable to read value from}

\item{lib}{name of js library used}

\item{package}{name where js library resides}
}
\description{
Use as Shiny output. First, use \code{renderHighcharts} in \code{server.R} to assign
the chart object to an Shiny output. Then create an chartOutput with the same name in #'
\code{ui.R}.
}
\author{
Thomas Reinholdsson, Ramnath Vaidyanathan
}

